refactor(cli-engine): dissolve src/definition into flat concept modules
Operator review: 'definition' discriminates nothing — in a library
everything is a definition; the folder actually meant 'types with no
behavior', a mechanism split. The ten concept modules (args, commands,
command-family, config-section, context, envelopes, events,
presentation, runtime, streams) now sit flat in src/ beside protocol
and config-loader, mirroring the draft's own organization. execution/
remains: it names the real internal subsystem.
