+++ packages/cli-engine/src/execution/events.ts
@@ -1,5 +1,5 @@
-import type { StreamEvent } from "../definition/envelopes";
-import type { EngineEvent, LogLevel, Severity } from "../definition/events";
+import type { StreamEvent } from "../envelopes";
+import type { EngineEvent, LogLevel, Severity } from "../events";
 import type { Invocation } from "./invocation";
 import { renderEventHuman } from "./render";
